Split UnifiedPushMessage to hold internal information
