+++ push/model/src/test/resources/message-format.json
@@ -1,6 +1,4 @@
 {
-  "ipAddress": null,
-  "clientIdentifier": null,
   "message": {
     "alert": "HELLO!",
     "sound": "default",
