+++ push/model/src/test/resources/message-tostrippedjson.json
@@ -1,6 +1,4 @@
 {
-  "ipAddress": null,
-  "clientIdentifier": null,
   "alert": "HELLO!",
   "criteria": {
     "categories": null,
